sync API code from upstream
